module7-task1: + selectors for the offers
