+++ src/store/selectors/selectors.js
@@ -1,0 +1,1 @@
+export * from './places-data/places-data';
